fix(rocm-kernels): key a custom module on its source, not just its name
`get_or_load_custom` keyed the in-memory map on the module name alone
while the disk cache keyed on the source, so the first source seen
under a name won and every revision of it silently resolved against
the module it was meant to replace. The README and
`RocmDevice::get_or_load_custom_func` both claimed the source was part
of the key -- true only from the second process onwards.

`ModuleName::Custom` now carries a SHA-256 of the source. The digest
is of the source alone, not the disk cache's `module_key`, which also
covers the staged header set and is far larger than most sources; it
is kept full width rather than truncated because a collision here is a
wrong-kernel launch, not a stale file. `BuiltIn` keeps its
pointer-thin key and hashes nothing.

Modules are never unloaded and cannot be: `function()` returns a bare
`hipFunction_t` that borrows nothing, so evicting a module dangles
every handle already resolved from it. Documented as an invariant --
callers generating a module per shape must bound their source set --
rather than papered over with an LRU. The compile gate has no such
constraint and is dropped once its module is resident.

`compile/mod.rs` was over the 400-line cap, so the toolchain seam
moves to `compile/hipcc.rs` and its tests to `compile/tests.rs`.
